move agent/server logic to pkg
